+++ src/main/java/lotto/LottoResults.java
@@ -1,14 +1,9 @@
 package lotto;
 
-import java.util.Arrays;
-import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class LottoResults {
-	private static final String MESSAGE = "%s개 일치 (%s원)- %s개";
-	private static final String RATE_MESSAGE = "총 수익률은 %1$,.2f 입니다.";
-
 	private final List<LottoResult> results;
 
 	public LottoResults(List<LottoResult> results) {
@@ -21,19 +16,7 @@ public class LottoResults {
 			.count());
 	}
 
-	public String getResultMessage() {
-		return Arrays.stream(Rank.values())
-			.sorted(Comparator.comparingInt(Rank::getMatchesCount))
-			.filter(target -> target.getAmountMoney() > 0)
-			.map(target -> String.format(MESSAGE, target.getMatchesCount(), target.getAmountMoney(), askCountOfRank(target)))
-			.collect(Collectors.joining("\n"));
-	}
-
-	public String getRateMessage(int buyMoney) {
-		Integer sumMoney = this.results.stream()
-			.map(LottoResult::getRankAmountMoney)
-			.reduce(0, Integer::sum);
-
-		return String.format(RATE_MESSAGE, (double) sumMoney / buyMoney);
+	public Stream<LottoResult> getLottoResults() {
+		return this.results.stream();
 	}
 }
