+++ src/main/java/lotto/Rank.java
@@ -1,5 +1,7 @@
 package lotto;
 
+import java.util.Arrays;
+import java.util.Comparator;
 import java.util.stream.Stream;
 
 public enum Rank {
@@ -18,10 +20,16 @@ public enum Rank {
 	}
 
 	public static Rank ofMatchesCount(int matchesCount) {
-		return Stream.of(Rank.values())
-			.filter(rank -> rank.getMatchesCount() == matchesCount)
+		return Stream.of(values())
+			.filter(target -> target.getMatchesCount() == matchesCount)
 			.findFirst()
 			.orElse(FIVE);
+	}
+
+	public static Stream<Rank> greaterThanCountZeroAndSortedAmount() {
+		return Arrays.stream(values())
+			.sorted(Comparator.comparingInt(Rank::getMatchesCount))
+			.filter(target -> target.getAmountMoney() > 0);
 	}
 
 	public int getMatchesCount() {
